+++ storage/ndb/test/ndbapi/testScanFilter.cpp
@@ -285,8 +285,8 @@ char op_string[] = "aoAO";
 * the six columns' name of test table
 */
 char col_string[] = "ijklmn";
-const int op_len = strlen(op_string);
-const int col_len = strlen(col_string);
+const int op_len = (int)strlen(op_string);
+const int col_len = (int)strlen(col_string);
 
 /*
 * get a random op from "aoAO"
@@ -369,7 +369,7 @@ int replace_a_to_str(char *source, int pos, char *newstr)
   BaseString::snprintf(temp+pos, strlen(newstr)+1, "%s", newstr);
   BaseString::snprintf(temp+pos+strlen(newstr), strlen(source)-pos, "%s", source+pos+1);
   BaseString::snprintf(source, strlen(temp)+1, "%s", temp);
-  return strlen(source);
+  return (int)strlen(source);
 }
 
 /*
@@ -496,9 +496,9 @@ void get_rand_op_str_compound(char *str)
   for(int i = 0; i < level; i++)
   {
     get_rand_op_str(small_str);
-    tmp = strlen(small_str);
+    tmp = (int)strlen(small_str);
     get_rand_op_str(small_str + tmp);   //get two operations
-    pos = get_rand_replace_pos(str, strlen(str));
+    pos = get_rand_replace_pos(str, (int)strlen(str));
     replace_a_to_str(str, pos, small_str);
   }
 
@@ -676,7 +676,7 @@ void check_all_tuples(char *str, bool *res)
 {    
   for (int i = 0; i < TUPLE_NUM; i++)
   {
-    if(check_one_tuple(i, str, strlen(str)))
+    if(check_one_tuple(i, str, (int)strlen(str)))
       res[i] = true;
   }
 }
